add README and clean up the code

+++ include/save.h
@@ -1,3 +1,12 @@
+/*
+ * @Description: 
+ * @Version: 1.0
+ * @Autor: bonbon
+ * @Date: 2021-08-01 19:58:22
+ * @LastEditors: bonbon
+ * @LastEditTime: 2021-08-02 13:56:52
+ */
+
 #pragma once 
 #include <iostream>
 #include <fstream>
@@ -13,7 +22,7 @@ void save_data(std::string path,
                float lidar_left_x,float lidar_left_y,
                float lidar_middle_x,float lidar_middle_y,
                float lidar_right_x,float lidar_right_y,
-               short x,short y,short yaw,
+               short x,short y,double yaw,
                int pot_x,int pot_y,double angle,unsigned char index
                );
 std::string date_init(void);
